move PoolDistr and ProtVer to cardano-ledger-core
