+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/PParams.hs
@@ -19,7 +19,6 @@ module Cardano.Ledger.Alonzo.PParams
   ( PParams' (..),
     PParams,
     emptyPParams,
-    ProtVer (..),
     PParamsUpdate,
     emptyPParamsUpdate,
     updatePParams,
@@ -30,6 +29,10 @@ module Cardano.Ledger.Alonzo.PParams
     extendPP,
     ppPParams,
     ppPParamsUpdate,
+    -- Deprecated
+    ProtVer,
+    pvMajor,
+    pvMinor,
   )
 where
 
@@ -60,6 +63,7 @@ import Cardano.Ledger.BaseTypes
     fromSMaybe,
     isSNothing,
   )
+import qualified Cardano.Ledger.BaseTypes as BT (ProtVer (..))
 import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Era
 import Cardano.Ledger.Pretty
@@ -82,7 +86,7 @@ import Cardano.Ledger.Serialization
     mapToCBOR,
   )
 import Cardano.Ledger.Shelley.Orphans ()
-import Cardano.Ledger.Shelley.PParams (HKD, ProtVer (..))
+import Cardano.Ledger.Shelley.PParams (HKD)
 import qualified Cardano.Ledger.Shelley.PParams as Shelley (PParams' (..))
 import Cardano.Ledger.Slot (EpochNo (..))
 import Control.DeepSeq (NFData)
@@ -147,7 +151,7 @@ data PParams' f era = PParams
     -- | Extra entropy
     _extraEntropy :: !(HKD f Nonce),
     -- | Protocol version
-    _protocolVersion :: !(HKD f ProtVer),
+    _protocolVersion :: !(HKD f BT.ProtVer),
     -- | Minimum Stake Pool Cost
     _minPoolCost :: !(HKD f Coin),
     -- new/updated for alonzo
@@ -291,7 +295,7 @@ emptyPParams =
       _tau = minBound,
       _d = minBound,
       _extraEntropy = NeutralNonce,
-      _protocolVersion = ProtVer 0 0,
+      _protocolVersion = BT.ProtVer 0 0,
       _minPoolCost = mempty,
       -- new/updated for alonzo
       _coinsPerUTxOWord = Coin 0,
@@ -629,3 +633,15 @@ ppPParamsUpdate (PParams feeA feeB mbb mtx mbh kd pd em no a0 rho tau d ex pv mp
 
 instance PrettyA (PParams' StrictMaybe era) where
   prettyA = ppPParamsUpdate
+
+{-# DEPRECATED ProtVer "Import from Cardano.Ledger.BaseTypes instead" #-}
+
+type ProtVer = BT.ProtVer
+
+{-# DEPRECATED pvMajor "Import from Cardano.Ledger.BaseTypes instead" #-}
+pvMajor :: ProtVer -> Natural
+pvMajor = BT.pvMajor
+
+{-# DEPRECATED pvMinor "Import from Cardano.Ledger.BaseTypes instead" #-}
+pvMinor :: ProtVer -> Natural
+pvMinor = BT.pvMinor
