+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/RewardUpdate.hs
@@ -21,14 +21,7 @@ import Cardano.Binary
     ToCBOR (..),
     encodeListLen,
   )
-import Cardano.Ledger.BaseTypes
-  ( ActiveSlotCoeff,
-    NonNegativeInterval,
-    ShelleyBase,
-    UnitInterval,
-    boundedRationalFromCBOR,
-    boundedRationalToCBOR,
-  )
+import Cardano.Ledger.BaseTypes (ActiveSlotCoeff, NonNegativeInterval, ProtVer (..), ShelleyBase, UnitInterval, boundedRationalFromCBOR, boundedRationalToCBOR)
 import Cardano.Ledger.Coin (Coin (..), DeltaCoin (..))
 import Cardano.Ledger.Credential (Credential (..))
 import qualified Cardano.Ledger.Crypto as CC (Crypto)
@@ -39,7 +32,6 @@ import Cardano.Ledger.Shelley.EpochBoundary
     Stake (..),
     poolStake,
   )
-import Cardano.Ledger.Shelley.PParams (ProtVer (..))
 import Cardano.Ledger.Shelley.RewardProvenance (RewardProvenancePool (..))
 import qualified Cardano.Ledger.Shelley.RewardProvenance as RP
 import Cardano.Ledger.Shelley.Rewards
