+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rewards.hs
@@ -48,14 +48,7 @@ import Cardano.Binary
     encodeListLen,
     encodeWord,
   )
-import Cardano.Ledger.BaseTypes
-  ( ActiveSlotCoeff,
-    BoundedRational (..),
-    NonNegativeInterval,
-    UnitInterval,
-    activeSlotVal,
-    invalidKey,
-  )
+import Cardano.Ledger.BaseTypes (ActiveSlotCoeff, BoundedRational (..), NonNegativeInterval, ProtVer, UnitInterval, activeSlotVal, invalidKey)
 import Cardano.Ledger.Coin
   ( Coin (..),
     coinToRational,
@@ -76,7 +69,6 @@ import Cardano.Ledger.Shelley.EpochBoundary
     maxPool',
   )
 import qualified Cardano.Ledger.Shelley.HardForks as HardForks
-import Cardano.Ledger.Shelley.PParams (ProtVer)
 import Cardano.Ledger.Shelley.RewardProvenance (RewardProvenancePool (..))
 import Cardano.Ledger.Shelley.TxBody (PoolParams (..), getRwdCred)
 import Cardano.Ledger.Val ((<->))
