+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Pool.hs
@@ -26,15 +26,7 @@ import Cardano.Binary
     encodeListLen,
   )
 import Cardano.Crypto.Hash.Class (sizeHash)
-import Cardano.Ledger.BaseTypes
-  ( Globals (..),
-    Network,
-    ShelleyBase,
-    StrictMaybe (..),
-    epochInfo,
-    invalidKey,
-    networkId,
-  )
+import Cardano.Ledger.BaseTypes (Globals (..), Network, ProtVer, ShelleyBase, StrictMaybe (..), epochInfo, invalidKey, networkId)
 import Cardano.Ledger.Coin (Coin)
 import qualified Cardano.Ledger.Core as Core
 import qualified Cardano.Ledger.Crypto as CC (Crypto (HASH))
@@ -43,7 +35,6 @@ import Cardano.Ledger.Keys (KeyHash (..), KeyRole (..))
 import Cardano.Ledger.Serialization (decodeRecordSum)
 import qualified Cardano.Ledger.Shelley.HardForks as HardForks
 import Cardano.Ledger.Shelley.LedgerState (PState (..))
-import Cardano.Ledger.Shelley.PParams (ProtVer)
 import qualified Cardano.Ledger.Shelley.SoftForks as SoftForks
 import Cardano.Ledger.Shelley.TxBody
   ( DCert (..),
