+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Rupd.hs
@@ -22,18 +22,7 @@ module Cardano.Ledger.Shelley.Rules.Rupd
   )
 where
 
-import Cardano.Ledger.BaseTypes
-  ( BlocksMade,
-    NonNegativeInterval,
-    ShelleyBase,
-    StrictMaybe (..),
-    UnitInterval,
-    activeSlotCoeff,
-    epochInfo,
-    maxLovelaceSupply,
-    randomnessStabilisationWindow,
-    securityParameter,
-  )
+import Cardano.Ledger.BaseTypes (BlocksMade, NonNegativeInterval, ProtVer, ShelleyBase, StrictMaybe (..), UnitInterval, activeSlotCoeff, epochInfo, maxLovelaceSupply, randomnessStabilisationWindow, securityParameter)
 import Cardano.Ledger.Coin (Coin (..))
 import qualified Cardano.Ledger.Core as Core
 import Cardano.Ledger.Era (Crypto, Era)
@@ -45,7 +34,6 @@ import Cardano.Ledger.Shelley.LedgerState
     pulseStep,
     startStep,
   )
-import Cardano.Ledger.Shelley.PParams (ProtVer)
 import Cardano.Ledger.Slot
   ( Duration (..),
     SlotNo,
